ci: narrow Windows payload validation
