+++ e2e/tests/scripts/scopes.test.ts
@@ -60,7 +60,19 @@ describe("workflow scope planner", () => {
     });
     expect(plan("pr", ["apps/desktop/src/main.ts"])).toMatchObject({
       scopes: { tools_dev_tests_required: true, tools_pack_tests_required: true },
+      enabled: { windows_tools_pack_payload_tests: false, ui_p0: false, playwright_critical: false },
+    });
+    expect(plan("pr", ["tools/pack/src/win/payload.ts"])).toMatchObject({
+      scopes: { tools_pack_tests_required: true, windows_tools_pack_payload_tests_required: true },
       enabled: { windows_tools_pack_payload_tests: true, ui_p0: false, playwright_critical: false },
+    });
+    expect(plan("pr", ["packages/launcher-proto/src/index.ts"])).toMatchObject({
+      scopes: { windows_tools_pack_payload_tests_required: true },
+      enabled: { windows_tools_pack_payload_tests: true },
+    });
+    expect(plan("pr", ["pnpm-lock.yaml"])).toMatchObject({
+      scopes: { windows_tools_pack_payload_tests_required: true },
+      enabled: { windows_tools_pack_payload_tests: true },
     });
     expect(plan("pr", ["docs/spec.md"])).toMatchObject({
       scopes: { workspace_validation_required: false },
@@ -98,12 +110,24 @@ describe("workflow scope planner", () => {
       trace: { escalations: [] },
     });
     expect(plan("merge-queue", ["apps/desktop/src/main.ts"])).toMatchObject({
+      enabled: { windows_tools_pack_payload_tests: false, workspace_unit_tests: true, e2e_vitest: false },
+      trace: { escalations: [] },
+    });
+    expect(plan("merge-queue", ["tools/pack/src/win/custom-installer.ts"])).toMatchObject({
       enabled: { windows_tools_pack_payload_tests: true, workspace_unit_tests: true, e2e_vitest: false },
       trace: { escalations: [] },
     });
     const medium = plan("merge-queue", ["apps/web/src/App.tsx"]);
     expect(medium.trace.escalations).toHaveLength(1);
     expect(Object.values(medium.scopes).filter((value) => typeof value === "boolean")).not.toContain(false);
+
+    const unknown = plan("merge-queue", ["some-new-root/file.ts"]);
+    expect(unknown).toMatchObject({ enabled: { windows_tools_pack_payload_tests: true } });
+    expect(unknown.trace.escalations).toHaveLength(1);
+  });
+
+  test("keeps the Windows payload workload in forced-full plans", () => {
+    expect(plan("full")).toMatchObject({ enabled: { windows_tools_pack_payload_tests: true } });
   });
 
   test("preserves the four-domain runtime-definition shadow candidate", () => {
